WIP: authenticator refactor
